requirements.txt
--find-links https://download.pytorch.org/whl/cu121
torch==2.2.0+cu121
torchvision==0.17.0+cu121
torchaudio==2.2.0+cu121
transformers==4.37.2
tokenizers==0.15.2
opencv-python==4.9.0.80
pillow==10.2.0
tqdm==4.66.1
numpy==1.24.3
accelerate==0.26.1
safetensors==0.4.1
huggingface-hub>=0.20.3
packaging>=23.2 
maplotlib==3.8.0
